requirements.txt
torch
torchvision
pytorch-lightning
tensorboardX
transformers
diffusers
opencv-python
albumentations
torchmetrics
mamba-ssm
PyYAML==6.0


